+++ packages/tempo/src/support/support.register.ts
@@ -1,8 +1,7 @@
 import { clearCache } from '#library/function.library.js';
 import { isEqual } from '#library/object.library.js';
-import { isDefined, isObject, isPlainObject, isSymbol, isUndefined } from '#library/assertion.library.js';
-import { ownKeys } from '#library/primitive.library.js';
-import { unwrap } from '#library/primitive.library.js';
+import { isDefined, isPlainObject, isSymbol, isUndefined } from '#library/assertion.library.js';
+import { ownKeys, unwrap } from '#library/primitive.library.js';
 import { sym } from './support.symbol.js';
 import type { Property } from '#library/type.library.js';